+++ src/test/groovy/ru/vyarus/gradle/plugin/python/cmd/AbstractCliMockSupport.groovy
@@ -26,6 +26,8 @@ abstract class AbstractCliMockSupport extends Specification {
 
     Project project
     TestLogger logger
+    private boolean execMocked
+    Map<Closure<String>, String> execCases = [:]
 
     boolean isWin = Os.isFamily(Os.FAMILY_WINDOWS)
 
@@ -36,21 +38,43 @@ abstract class AbstractCliMockSupport extends Specification {
     void setup() {
         project = Stub(Project)
         logger = new TestLogger()
-        project.getLogger()  >>  { logger }
+        project.getLogger() >> { logger }
         project.getProjectDir() >> { dir.root }
         project.file(_) >> { new File(dir.root, it[0]) }
     }
 
+    // use to provide specialized output for executed commands
+    // (e.g. under pip tests to cactch python virtualenv detection)
+    // closure accepts called command line
+    void execCase(Closure<String> closure, String output) {
+        execCases.put(closure, output)
+    }
+
     void mockExec(Project project, String output, int res) {
+        assert !execMocked, "Exec can be mocked just once!"
         // check execution with logs without actual execution
-        project.exec(_) >> {Closure spec ->
+        project.exec(_) >> { Closure spec ->
             DefaultExecAction act = ConfigureUtil.configure(spec,
                     new DefaultExecAction(Stub(PathToFileResolver),
                             Executors.newSingleThreadExecutor(),
                             new DefaultBuildCancellationToken()))
+
+            String cmd = "${act.executable} ${act.args.join(' ')}"
+            println ">> Mocked exec: $cmd"
+            String out = output
+            execCases.each { k, v ->
+                if (k.call(cmd)) {
+                    println ">> Special exec case detected, output become: $v"
+                    out = v
+                }
+            }
+            if (out==output) {
+                println ">> Default execution, output: $out"
+            }
+
             OutputStream os = act.standardOutput
-            if (output) {
-                os.write(output.bytes)
+            if (out) {
+                os.write(out.bytes)
             }
             return new ExecRes(res)
         }